Move Screens{CoordinatorLayout,Animation} to separate package
